// src/types/User.d.ts
export type User = {
  createdAt: string;
  updatedAt: string;
  id: string;
  walletAddress: string;
  email: string;
  lastName: string;
  firstName: string;
  userName: string;
  profileIcon: string;
  profileImage: string;
  fishingCareer: number;
  mainField: string;
  mainRod: string;
  mainReel: string;
  mainLine: string;
  introduction: string;
  role: string;
  isGuest: boolean;
};